useState and NavBar
